requirements.txt
click==7.1.2
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap4==4.0.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
visitor==0.1.3
Werkzeug==1.0.1
